\name{BMI713pack-package}
\alias{BMI713pack-package}
\alias{BMI713pack}
%"compar"    "ensViz"    "impath"    "myst"      "plotImage"
\alias{compar}
\alias{impath}
\alias{myst}
\alias{plotImage}
\alias{vcImages}
\docType{package}
\title{
\packageTitle{BMI713pack}
}
\description{
\packageDescription{BMI713pack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BMI713pack}
\packageIndices{BMI713pack}

This package includes software support for the 2016
edition of BMI 713 at Harvard Medical School.

\code{compar} helps use \code{\link[microbenchmark]{microbenchmark}}
to report on comparative performance of two functions

\code{myst} is a mystery function to be interpreted

\code{impath} is used to render a display of the mouse otocyst

\code{plotImage} is used to render a display from a publication

\code{vcImages} is used to retrieve png files for rendering in lecture slides
}
\author{
\packageAuthor{BMI713pack}

Maintainer: \packageMaintainer{BMI713pack}
}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
\examples{
ensViz(ideo=FALSE)
}
